recommender/pyproject.toml
[project]
name = "recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "findspark>=2.0.1",
    "joblib>=1.4.2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyspark>=3.5.5",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "setuptools>=76.1.0",
    "spotipy>=2.25.1",
    "streamlit>=1.43.2",
]
